refactor: remove redundant props files, beautify imports, create index files in components and pages
