+++ components/src/App.tsx
@@ -1,6 +1,7 @@
 import { Component } from 'react';
 import React from 'react';
-import { Router } from './Router';
+
+import { Router } from './AppRouter';
 
 export class App extends Component {
   render() {
