+++ components/src/Pages/About/About.module.scss
@@ -1,4 +1,3 @@
 .about {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
   font-size: 3rem;
 }
