+++ components/src/Pages/About/About.tsx
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+
 import styles from './About.module.scss';
 
 export class About extends Component {
