+++ components/src/Pages/index.ts
@@ -1,0 +1,3 @@
+export * from './404/NotFound';
+export * from './About/About';
+export * from './Home/Home';
