+++ components/src/components/Card/Card.module.scss
@@ -1,5 +1,4 @@
 .card {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
   display: flex;
   flex-direction: column;
   gap: 0.5rem;
