+++ components/src/components/SearchBar/SearchBar.interface.ts
@@ -1,0 +1,7 @@
+export interface SearchBarProps {
+  id?: number;
+}
+
+export interface SearchState {
+  searchValue: string;
+}
