+++ components/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './SearchBar/SearchBar';
+export * from './Products/Products';
+export * from './Card/Card';
