+++ components/src/enums/RoutePaths.ts
@@ -1,0 +1,5 @@
+export enum RoutePaths {
+  ROOT = '/',
+  ABOUT = '/about',
+  ALL = '*',
+}
