+++ components/src/layout/Footer/Footer.module.scss
@@ -4,7 +4,6 @@
   background-color: aqua;
   align-items: center;
   gap: 0.5rem 2rem;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
   flex-wrap: wrap;
   color: #444;
   font-weight: 500;
