+++ components/src/layout/Header/Header.module.scss
@@ -3,7 +3,6 @@
 }
 
 .header {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
   padding: 2rem 0 1rem;
 }
