--- components/src/layout/Header/Header.props.ts
@@ -1,4 +1,0 @@
-import { DetailedHTMLProps, HTMLAttributes } from 'react';
-
-export interface HeaderProps
-  extends DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement> {}
