+++ components/src/layout/Header/Header.tsx
@@ -1,17 +1,19 @@
-import styles from './Header.module.scss';
 import { NavLink } from 'react-router-dom';
 import { Component } from 'react';
+
+import styles from './Header.module.scss';
+
+import { menus } from './menus';
 
 export class Header extends Component {
   render() {
     return (
       <header className={styles.header}>
-        <NavLink className={styles.link} to="/">
-          Home
-        </NavLink>
-        <NavLink className={styles.link} to="/about">
-          About
-        </NavLink>
+        {menus.map((elem) => (
+          <NavLink to={elem.to} className={styles.link} key={elem.id}>
+            {elem.text}
+          </NavLink>
+        ))}
       </header>
     );
   }
