+++ components/src/layout/Header/menus.ts
@@ -1,0 +1,14 @@
+import { RoutePaths } from '../../enums/RoutePaths';
+
+export const menus = [
+  {
+    to: RoutePaths.ROOT,
+    text: 'Home',
+    id: 1,
+  },
+  {
+    to: RoutePaths.ABOUT,
+    text: 'About',
+    id: 2,
+  },
+];
