+++ components/src/layout/Layout.module.scss
@@ -1,3 +1,7 @@
+* {
+  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+}
+
 .wrapper {
   width: min(1170px, 100%);
   margin: 0 auto;
