+++ components/src/layout/Layout.tsx
@@ -1,8 +1,10 @@
-import styles from './Layout.module.scss';
-import { Header } from './Header/Header';
-import { Footer } from './Footer/Footer';
 import React, { Component } from 'react';
 import { Outlet } from 'react-router-dom';
+
+import { Header } from './Header/Header';
+import { Footer } from './Footer/Footer';
+
+import styles from './Layout.module.scss';
 
 export class Layout extends Component {
   render() {
