+++ components/vite.config.ts
@@ -3,7 +3,6 @@ import react from '@vitejs/plugin-react';
 import svgr from 'vite-plugin-svgr';
 import eslintPlugin from 'vite-plugin-eslint';
 
-// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [svgr(), react(), eslintPlugin()],
+  plugins: [svgr(), react()],
 });
